Update autoconfig.settings.sample.php
When reduced to a single domain PHP will error because of unexpected ] in line 19 and the base file is annoyed at line 15 of the missing array. $cfg->domains = array( 'example.com', 'example.net', 'example.org' ); and $cfg->domains = array( 'example.com' ); are valid arrays.
Thx for the Project, works like a charm.